+++ console/config/console.php
@@ -4,13 +4,13 @@ return [
     'basePath' => dirname(__DIR__),
     'controllerNamespace' => 'console\controllers',
     'controllerMap'=>[
+        'message'=>[
+            'class'=>'console\controllers\ExtendedMessageController'
+        ],
         'migrate'=>[
             'class'=>'yii\console\controllers\MigrateController',
             'migrationPath'=>'@common/migrations',
             'migrationTable'=>'{{%system_migration}}'
-        ],
-        'message-migrate'=>[
-            'class'=>'console\controllers\MessageMigrateController'
         ],
         'rbac'=>[
             'class'=>'console\controllers\RbacController'